move file discovery out of utils
